+++ build.gradle
@@ -5,7 +5,7 @@ plugins {
 }
 
 group 'com.sayurbox.kale'
-version '1.0.0'
+version '1.5.0'
 
 apply plugin: 'jacoco'
 
@@ -28,7 +28,6 @@ jar {
 }
 
 dependencies {
-    implementation "com.netflix.hystrix:hystrix-core:1.5.18"
     implementation "org.slf4j:slf4j-api:1.7.25"
     implementation "com.squareup.okhttp3:okhttp:3.14.6"
     implementation "com.squareup.okhttp3:logging-interceptor:3.14.6"
