--- src/main/java/com/sayurbox/kale/api/GetAllocateDataResponse.java
@@ -1,7 +1,0 @@
-package com.sayurbox.kale.api;
-
-import com.sayurbox.kale.common.client.DataResponse;
-import com.sayurbox.kale.featureflag.client.GetAllocateResponse;
-
-public class GetAllocateDataResponse extends DataResponse<GetAllocateResponse> {
-}
