+++ src/main/java/com/sayurbox/kale/common/command/KaleCommand.java
@@ -1,45 +1,51 @@
 package com.sayurbox.kale.common.command;
 
 import com.google.gson.Gson;
-import com.netflix.hystrix.HystrixCommand;
-import com.netflix.hystrix.HystrixCommandProperties;
-import com.sayurbox.kale.config.KaleHystrixParams;
+import io.github.resilience4j.circuitbreaker.CallNotPermittedException;
+import io.github.resilience4j.circuitbreaker.CircuitBreaker;
+import io.github.resilience4j.decorators.Decorators;
 import okhttp3.OkHttpClient;
 import okhttp3.Request;
 import okhttp3.Response;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.UnsupportedEncodingException;
+import java.io.IOException;
+import java.util.Collections;
+import java.util.function.Supplier;
 
-public abstract class KaleCommand<T> extends HystrixCommand<T> {
+// Helper for http client wrapped by circuit breaker
+public abstract class KaleCommand<T> {
 
     private static final Logger logger = LoggerFactory.getLogger(KaleCommand.class);
 
     private final OkHttpClient okHttpClient;
+    private final CircuitBreaker circuitBreaker;
+    private final boolean isCircuitBreakerEnabled;
     protected final Gson gson;
     protected final String baseUrl;
 
-    protected KaleCommand(KaleHystrixParams hystrixParams,
+    protected KaleCommand(CircuitBreaker circuitBreaker,
                           OkHttpClient okHttpClient,
+                          boolean isCircuitBreakerEnabled,
                           String baseUrl) {
-        super(hystrixSetter(hystrixParams));
         this.okHttpClient = okHttpClient;
+        this.circuitBreaker = circuitBreaker;
+        this.isCircuitBreakerEnabled = isCircuitBreakerEnabled;
         this.gson = new Gson();
         this.baseUrl = baseUrl;
     }
 
-    @Override
-    public T run() {
+    protected T run() {
         Response response = null;
         try {
             Request request = createRequest();
-            logger.info("requesting kale: " + request.url());
+            logger.info("requesting kale: {}", request.url());
             response = okHttpClient.newCall(request).execute();
             return handleResponse(response);
         } catch (Exception e) {
             logger.error("failed request from kale, ", e);
-            throw new RuntimeException(e);
+            return getFallback();
         } finally {
             if (response != null && response.body() != null) {
                 response.body().close();
@@ -47,26 +53,28 @@ public abstract class KaleCommand<T> extends HystrixCommand<T> {
         }
     }
 
+    public T execute() {
+        // ref code examples:
+        // https://github.com/thombergs/code-examples/blob/master/resilience4j/timelimiter/src/main/java/io/reflectoring/resilience4j/timelimiter/Examples.java
+        Supplier<T> supplier = this::run;
+        if (!this.isCircuitBreakerEnabled) {
+            return supplier.get();
+        }
+        Decorators.DecorateSupplier<T> decorated = Decorators
+                .ofSupplier(supplier)
+                .withCircuitBreaker(this.circuitBreaker)
+                .withFallback(Collections.singletonList(CallNotPermittedException.class), e -> {
+                    logger.warn("execute fallback CallNotPermittedException");
+                    return getFallback();
+                });
+        Supplier<T> decorator = decorated.decorate();
+        return decorator.get();
+    }
 
-    @Override
     public abstract T getFallback();
 
-    protected abstract T handleResponse(Response response) throws Exception;
+    protected abstract T handleResponse(Response response) throws IOException;
 
-    protected abstract Request createRequest() throws UnsupportedEncodingException;
-
-    private static Setter hystrixSetter(KaleHystrixParams params) {
-        HystrixCommand.Setter config = HystrixCommand.Setter.withGroupKey(params.getHystrixCommandGroupKey());
-        HystrixCommandProperties.Setter properties = HystrixCommandProperties.Setter();
-        properties.withExecutionTimeoutInMilliseconds(params.getExecutionTimeout());
-        properties.withCircuitBreakerSleepWindowInMilliseconds(params.getCircuitBreakerSleepWindow());
-        properties.withExecutionIsolationStrategy(HystrixCommandProperties.ExecutionIsolationStrategy.THREAD);
-        properties.withCircuitBreakerEnabled(true);
-        properties.withCircuitBreakerRequestVolumeThreshold(params.getCircuitBreakerRequestVolumeThreshold());
-        properties.withMetricsRollingStatisticalWindowInMilliseconds(params.getMetricRollingStatisticalWindow());
-        properties.withMetricsHealthSnapshotIntervalInMilliseconds(params.getMetricsHealthSnapshotInterval());
-        config.andCommandPropertiesDefaults(properties);
-        return config;
-    }
+    protected abstract Request createRequest();
 
 }
