+++ src/main/java/com/sayurbox/kale/config/KaleConfig.java
@@ -7,23 +7,16 @@ public class KaleConfig {
     private final String baseUrl;
     private final Long executionTimeout;
     private final boolean logEnabled;
-    private final KaleHystrixParams hystrixParams;
     private final CircuitBreakerParams circuitBreakerParams;
 
     private KaleConfig(String baseUrl,
                        Long executionTimeout,
                        CircuitBreakerParams circuitBreakerParams,
-                       KaleHystrixParams hystrixParams,
                        boolean logEnabled) {
         this.baseUrl = requireNonNull(baseUrl);
         this.circuitBreakerParams = circuitBreakerParams;
         this.executionTimeout = executionTimeout;
-        this.hystrixParams = hystrixParams;
         this.logEnabled = logEnabled;
-    }
-
-    public KaleHystrixParams getHystrixParams() {
-        return hystrixParams;
     }
 
     public CircuitBreakerParams getCircuitBreakerParams() {
@@ -45,13 +38,6 @@ public class KaleConfig {
     public static class Builder {
 
         private String baseUrl;
-        // hystrix will be deprecated
-        private Integer hystrixExecutionTimeout = 5000;
-        private Integer hystrixCircuitBreakerSleepWindow = 1000;
-        private Integer hystrixCircuitBreakerRequestVolumeThreshold = 10;
-        private Integer hystrixRollingStatisticalWindow = 10000;
-        private Integer hystrixHealthSnapshotInterval = 500;
-
         private boolean logEnabled = false;
         private Integer executionTimeout = 5000;
         private boolean circuitBreakerEnabled = true;
@@ -64,32 +50,6 @@ public class KaleConfig {
 
         public Builder withBaseUrl(String url) {
             this.baseUrl = url;
-            return this;
-        }
-
-        public Builder withHystrixExecutionTimeout(int hystrixExecutionTimeout) {
-            this.hystrixExecutionTimeout = hystrixExecutionTimeout;
-            return this;
-        }
-
-        public Builder withHystrixCircuitBreakerSleepWindow(int hystrixCircuitBreakerSleepWindow) {
-            this.hystrixCircuitBreakerSleepWindow = hystrixCircuitBreakerSleepWindow;
-            return this;
-        }
-
-        public Builder withHystrixCircuitBreakerRequestVolumeThreshold(
-                int hystrixCircuitBreakerRequestVolumeThreshold) {
-            this.hystrixCircuitBreakerRequestVolumeThreshold = hystrixCircuitBreakerRequestVolumeThreshold;
-            return this;
-        }
-
-        public Builder withHystrixRollingStatisticalWindow(int hystrixRollingStatisticalWindow) {
-            this.hystrixRollingStatisticalWindow = hystrixRollingStatisticalWindow;
-            return this;
-        }
-
-        public Builder withHystrixHealthSnapshotInterval(int hystrixHealthSnapshotInterval) {
-            this.hystrixHealthSnapshotInterval = hystrixHealthSnapshotInterval;
             return this;
         }
 
@@ -124,11 +84,6 @@ public class KaleConfig {
         }
 
         public KaleConfig build() {
-            KaleHystrixParams hystrixParams = new KaleHystrixParams(hystrixExecutionTimeout,
-                    hystrixCircuitBreakerSleepWindow,
-                    hystrixCircuitBreakerRequestVolumeThreshold,
-                    hystrixRollingStatisticalWindow,
-                    hystrixHealthSnapshotInterval);
             CircuitBreakerParams circuitBreakerParams = new CircuitBreakerParams(
                     circuitBreakerEnabled,
                     failureVolumeThreshold,
@@ -137,7 +92,6 @@ public class KaleConfig {
             return new KaleConfig(baseUrl,
                     executionTimeout.longValue(),
                     circuitBreakerParams,
-                    hystrixParams,
                     logEnabled);
         }
     }
