--- src/main/java/com/sayurbox/kale/exception/KaleException.java
@@ -1,9 +1,0 @@
-package com.sayurbox.kale.exception;
-
-public class KaleException extends Exception {
-
-    public KaleException(String message) {
-        super(message);
-    }
-
-}
